Refactoring, TODO making

+++ DomKulturyBackend/src/main/java/com/example/demo/services/implementations/UserServiceImpl.java
@@ -1,31 +1,46 @@
 package com.example.demo.services.implementations;
 
 import com.example.demo.dto.UserDTO;
-import com.example.demo.exceptions.EmailInUseException;
-import com.example.demo.exceptions.UserNotFoundException;
-import com.example.demo.exceptions.UsernameTakenException;
+import com.example.demo.exceptions.*;
 import com.example.demo.mappers.UserMapper;
+import com.example.demo.models.ERole;
+import com.example.demo.models.Role;
 import com.example.demo.models.Subject;
 import com.example.demo.models.User;
+import com.example.demo.payload.request.PasswordChangeRequest;
 import com.example.demo.repository.UserRepository;
+import com.example.demo.services.interfaces.RoleService;
 import com.example.demo.services.interfaces.SubjectService;
 import com.example.demo.services.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
     private final UserMapper userMapper;
-
+    private final RoleService roleService;
+    private final PasswordEncoder encoder;
+    private final AuthenticationManager authenticationManager;
+    private final SubjectService subjectService;
 
     @Autowired
-    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
+    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper, RoleService roleService, PasswordEncoder encoder, AuthenticationManager authenticationManager, SubjectService subjectService) {
         this.userRepository = userRepository;
         this.userMapper = userMapper;
+        this.roleService = roleService;
+        this.encoder = encoder;
+        this.authenticationManager = authenticationManager;
+        this.subjectService = subjectService;
     }
 
     @Override
@@ -46,6 +61,11 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(Long id) {
         User user = findById(id);
+
+        if (subjectService.findAll().stream().filter(subject -> subject.getTeacher().getId() == id).findAny().isPresent()) {
+            throw new UserAssignedToSubjectException();
+        }
+
         userRepository.delete(user);
     }
 
@@ -54,7 +74,6 @@ public class UserServiceImpl implements UserService {
 
         User user = findById(id);
 
-        //TODO inny wyjątek, żeby się zgadzało
         if(!userDTO.getId().equals(id)) {
             throw new UserNotFoundException(id);
         }
@@ -80,4 +99,34 @@ public class UserServiceImpl implements UserService {
     public User save(User user) {
         return userRepository.save(user);
     }
+
+    @Override
+    public void changeRole(ERole newRole, Long id) {
+        Role role = roleService.findByName(newRole);
+        User user = findById(id);
+
+        Set<Role> newRoles = new HashSet<>();
+        newRoles.add(role);
+
+        user.setRoles(newRoles);
+        save(user);
+    }
+
+    @Override
+    public void changePassword(PasswordChangeRequest passwordChangeRequest, Long id) {
+        User user = findById(id);
+
+        Authentication authentication = null;
+
+        authentication = authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(user.getUsername(), passwordChangeRequest.getPassword()));
+
+        if (authentication == null) {
+            throw new WrongPasswordException();
+        }
+        user.setPassword(encoder.encode(passwordChangeRequest.getNewPassword()));
+        save(user);
+    }
+
+
 }
